cb2/requirements.txt
# Carbon Black API and SDK requirements
cbapi>=2.0.0
requests>=2.28.1
pyyaml>=6.0
python-dateutil>=2.8.2

# Web application
flask>=2.2.2
flask-cors>=3.0.10
flask-sqlalchemy>=3.0.3
flask-marshmallow>=0.14.0
marshmallow-sqlalchemy>=0.28.1

# Database
sqlalchemy>=2.0.0
sqlalchemy-utils>=0.40.0
psycopg2-binary>=2.9.5
flask-migrate>=4.0.0

# Frontend
python-dotenv>=1.0.0

# Security
bcrypt>=4.0.1
python-jose>=3.3.0 